+++ core/translate/emitter.rs
@@ -8,7 +8,6 @@ use limbo_sqlite3_parser::ast::{self, SortOrder};
 use tracing::{instrument, Level};
 
 use super::aggregation::emit_ungrouped_aggregation;
-use super::check_constraint::{check_col_is_referred, translate_check_constraint};
 use super::expr::{translate_condition_expr, translate_expr, ConditionMetadata};
 use super::group_by::{
     group_by_agg_phase, group_by_emit_row_phase, init_group_by, GroupByMetadata, GroupByRowSource,
@@ -26,6 +25,7 @@ use super::subquery::emit_subqueries;
 use crate::error::SQLITE_CONSTRAINT_PRIMARYKEY;
 use crate::function::Func;
 use crate::schema::{Index, IndexColumn, Schema};
+use crate::translate::check_constraint::new_translate_check_constraint;
 use crate::translate::plan::{DeletePlan, Plan, Search};
 use crate::translate::values::emit_values;
 use crate::util::exprs_are_equivalent;
@@ -1275,68 +1275,47 @@ fn emit_update_insns(
             program.preassign_label_to_next_insn(record_label);
         }
 
-        for check_constraint in btree_table.table_check_constraints.iter() {
-            let jump_if_true = program.allocate_label();
-            translate_check_constraint(
-                program,
-                &check_constraint.expr,
-                btree_table
-                    .columns
+        new_translate_check_constraint(program, &btree_table, start, &t_ctx.resolver);
+        /*
+                let check_constraints: Vec<_> = table_ref
+                    .columns()
                     .iter()
-                    .enumerate()
-                    .map(|(i, column)| (start + i, column))
-                    .collect::<Vec<_>>()
-                    .as_ref(),
-                Some(jump_if_true),
-                &t_ctx.resolver,
-            );
+                    .filter(|col| col.check_constraint.is_some())
+                    .map(|col| col.check_constraint.as_ref().unwrap())
+                    .collect();
+                for constraint in &check_constraints {
+                    for column in table_ref.columns().iter() {
+                        if !check_col_is_referred(constraint, column.name.as_ref().map_or("", |name| &name))
+                        {
+                            continue;
+                        }
+                        let jump_if_true = program.allocate_label();
+                        translate_check_constraint(
+                            program,
+                            constraint,
+                            table_ref
+                                .columns()
+                                .iter()
+                                .enumerate()
+                                .map(|(i, col)| (start + i, col))
+                                .collect::<Vec<_>>()
+                                .as_ref(),
+                            Some(jump_if_true),
+                            &t_ctx.resolver,
+                        );
 
-            use crate::error::SQLITE_CONSTRAINT_CHECK;
-            program.emit_insn(Insn::Halt {
-                err_code: SQLITE_CONSTRAINT_CHECK,
-                description: check_constraint.description(),
-            });
+                        use crate::error::SQLITE_CONSTRAINT_CHECK;
+                        let description = constraint.to_string();
+                        program.emit_insn(Insn::Halt {
+                            err_code: SQLITE_CONSTRAINT_CHECK,
+                            description: description.to_string(),
+                        });
 
-            program.preassign_label_to_next_insn(jump_if_true);
-        }
-        let check_constraints: Vec<_> = table_ref
-            .columns()
-            .iter()
-            .filter(|col| col.check_constraint.is_some())
-            .map(|col| col.check_constraint.as_ref().unwrap())
-            .collect();
-        for constraint in &check_constraints {
-            for column in table_ref.columns().iter() {
-                if !check_col_is_referred(constraint, column.name.as_ref().map_or("", |name| &name))
-                {
-                    continue;
+                        program.preassign_label_to_next_insn(jump_if_true);
+                        break;
+                    }
                 }
-                let jump_if_true = program.allocate_label();
-                translate_check_constraint(
-                    program,
-                    constraint,
-                    table_ref
-                        .columns()
-                        .iter()
-                        .enumerate()
-                        .map(|(i, col)| (start + i, col))
-                        .collect::<Vec<_>>()
-                        .as_ref(),
-                    Some(jump_if_true),
-                    &t_ctx.resolver,
-                );
-
-                use crate::error::SQLITE_CONSTRAINT_CHECK;
-                let description = constraint.to_string();
-                program.emit_insn(Insn::Halt {
-                    err_code: SQLITE_CONSTRAINT_CHECK,
-                    description: description.to_string(),
-                });
-
-                program.preassign_label_to_next_insn(jump_if_true);
-                break;
-            }
-        }
+        */
 
         let record_reg = program.alloc_register();
         program.emit_insn(Insn::MakeRecord {
